Move Arduino.h #include directive to TMC26XStepper.h
TMC26XStepper.h uses the non-standard boolean type which is defined in the Arduino core library. The use of this type in TMC26XStepper.h relied on Arduino.h having been included externally before TMC26XStepper.h. While this would typically be done automatically by the Arduino sketch preprocessor if TMC26XStepper.h was included from an .ino file, that behavior is not guaranteed and the Arduino.h #include directive is not automatically to any other type of source file. Thus the best solution is to simply include the external dependency in the file that requires it.